requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mypy==1.7.1
mypy-extensions==1.0.0
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
ruff==0.4.2
setuptools==69.5.1
sqlparse==0.5.0
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
